some changes
- now `./src/index.js` is main exports instead of usage example

- old `./src/index.js` is now `/index.js`

- more changes
