+++ .eslintrc.js
@@ -10,6 +10,7 @@ module.exports = {
 		'indent': [2, 'tab', { 'SwitchCase': 1 }],
 		'no-tabs': [0],
 		'no-use-before-define': [0],
+		'no-restricted-syntax': [0],
 		/* only because it used as if else with a clear syntax I created it's been silenced */
 		'no-nested-ternary': [0],
 		'no-mixed-operators': [0],
